backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kpix-backend"
version = "0.1.0"
description = "FastAPI backend for the kpix performance platform."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
  "fastapi>=0.110,<0.112",
  "uvicorn[standard]>=0.27,<0.29",
  "sqlalchemy[asyncio]>=2.0,<2.1",
  "asyncpg>=0.29",
  "alembic>=1.13,<2.0",
  "passlib[argon2]>=1.7,<2.0",
  "PyJWT>=2.8,<3.0",
  "pydantic>=2.6,<3.0",
  "pydantic-settings>=2.2,<3.0",
  "python-multipart>=0.0.9",
  "openpyxl>=3.1,<4.0",
  "email-validator>=2.1,<3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4,<8.0",
  "pytest-asyncio>=0.23,<0.25",
  "httpx>=0.27,<0.28",
  "anyio>=4.3,<5.0",
  "coverage>=7.4,<8.0",
]

[project.scripts]
kpix-backend = "kpix_backend.main:run"

[tool.hatch.build.targets.wheel]
packages = ["src/kpix_backend"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]
